src/backend-py/requirements.txt
# Django Core
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-environ==0.11.2

# Database
psycopg2-binary==2.9.10
redis==5.2.1
django-redis==5.4.0

# AI & ML
openai>=1.58.1,<2.0.0
anthropic==0.42.0
langchain==0.3.16
langchain-openai==0.3.0
langchain-community==0.3.16

# Vector Store
qdrant-client==1.12.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2
python-magic==0.4.27
Pillow==11.1.0

# Audio/Video Processing
ffmpeg-python==0.2.0
SpeechRecognition==3.11.0

# OCR
pytesseract==0.3.13

# WebSocket Support
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2

# Authentication & Security
djangorestframework-simplejwt==5.4.0
cryptography==44.0.0

# HTTP & Requests
requests==2.32.3
httpx==0.28.1

# Task Queue
celery==5.4.0
django-celery-beat==2.7.0
flower==2.0.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.5
python-multipart==0.0.20

# Development
black==24.10.0
flake8==7.1.1
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
